+++ waku/v2/protocol/store/waku_store_protocol_test.go
@@ -73,11 +73,11 @@ func TestWakuStoreProtocolNext(t *testing.T) {
 	topic1 := "1"
 	pubsubTopic1 := "topic1"
 
-	msg1 := tests.CreateWakuMessage(topic1, float64(1))
-	msg2 := tests.CreateWakuMessage(topic1, float64(2))
-	msg3 := tests.CreateWakuMessage(topic1, float64(3))
-	msg4 := tests.CreateWakuMessage(topic1, float64(4))
-	msg5 := tests.CreateWakuMessage(topic1, float64(5))
+	msg1 := tests.CreateWakuMessage(topic1, 1)
+	msg2 := tests.CreateWakuMessage(topic1, 2)
+	msg3 := tests.CreateWakuMessage(topic1, 3)
+	msg4 := tests.CreateWakuMessage(topic1, 4)
+	msg5 := tests.CreateWakuMessage(topic1, 5)
 
 	s1.MsgC <- protocol.NewEnvelope(msg1, pubsubTopic1)
 	s1.MsgC <- protocol.NewEnvelope(msg2, pubsubTopic1)
